added DomainServiceProvider

+++ src/Api/Application/Car/Command/Handlers/DeleteCarHandler.php
@@ -2,26 +2,36 @@
 
 namespace Api\Application\Car\Command\Handlers;
 
+use Api\Domain\Car\Exception\CarNotAvailableDeleteException;
+use Api\Domain\Car\Model\Car;
 use Api\Domain\Car\Repository\CarRepository;
+use Api\Domain\Car\Service\CarService;
 use Api\Application\Car\Command\Commands\DeleteCarCommand;
 use Api\Domain\Car\Exception\CarNotFoundException;
 
 class DeleteCarHandler
 {
     public function __construct(
-        private CarRepository $repository
+        private CarRepository $repository,
+        private CarService $service
     ) {
 
     }
 
     /**
-     * @throws CarNotFoundException
+     * @throws CarNotFoundException|CarNotAvailableDeleteException
      */
     public function handle(DeleteCarCommand $command): void
     {
+        /** @var Car|null $car */
         if (!$car = $this->repository->findOneById($command->getId())) {
             throw new CarNotFoundException();
         }
+
+        if (!$this->service->isAvailableDelete($car->id)) {
+            throw new CarNotAvailableDeleteException();
+        }
+
         $this->repository->delete($car);
     }
 }
